feat(bpdm): fix search legal-entity
Refs: CPLP-2976

+++ src/externalsystems/Bpdm.Library/Models/BpdmAddressDto.cs
@@ -92,7 +92,7 @@ public record BpdmPostalDeliveryPointDto(
 );
 
 public record BpdmGeographicCoordinatesDto(
-    int Longitude,
-    int Latitude,
-    int Altitude
+    double Longitude,
+    double Latitude,
+    double Altitude
 );
